// client/web/types/generated/HighlightReason.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type HighlightReason = { "BoostedCutoff": { kills: number, } } | { "TrapKill": { kills: number, } } | { "Demolition": { kills: number, } } | { "GoalRun": { points: number, } } | { "ComboFrenzy": { max_chain: number, } } | { "FeedingFrenzy": { pickups: number, } };